requirements.txt
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
crypto==1.4.1
cryptography==2.4.2
future==0.17.1
gevent==1.3.4
greenlet==0.4.13
idna==2.7
multidict==4.5.2
Naked==0.1.31
Pillow==5.3.0
pyasn1==0.4.4
pycparser==2.18
pycryptodome==3.7.2
pycryptodomex==3.7.2
PyGithub==1.40
PyJWT==1.6.4
PyMySQL==0.9.2
pyTelegramBotAPI==3.6.6
python-alipay-sdk==1.8.0
pytz==2018.7
PyYAML==3.13
qrcode==6.0
requests==2.19.1
rsa==4.0
shellescape==3.4.1
six==1.12.0
urllib3==1.23
yarl==1.3.0
